requirements.txt
flet~=0.22.1
g4f~=0.3.1.0
openai-whisper